requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
kafka-python-ng==2.2.3
magic-filter==1.0.12
multidict==6.1.0
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
propcache==0.3.0
py4j==0.10.9.7
pydantic==2.10.6
pydantic_core==2.27.2
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
